Refactor new report page client logic
